+++ actionview/lib/action_view/lookup_context.rb
@@ -55,9 +55,7 @@ module ActionView
 
     class DetailsKey #:nodoc:
       alias :eql? :equal?
-      alias :object_hash :hash
 
-      attr_reader :hash
       @details_keys = Concurrent::Map.new
 
       def self.get(details)
@@ -72,8 +70,16 @@ module ActionView
         @details_keys.clear
       end
 
+      def self.empty?; @details_keys.empty?; end
+
+      def self.digest_caches
+        @details_keys.values.map(&:digest_cache)
+      end
+
+      attr_reader :digest_cache
+
       def initialize
-        @hash = object_hash
+        @digest_cache = Concurrent::Map.new
       end
     end
 
@@ -198,6 +204,10 @@ module ActionView
 
       @details = initialize_details({}, details)
       self.view_paths = view_paths
+    end
+
+    def digest_cache
+      details_key.digest_cache
     end
 
     def initialize_details(target, details)