+++ railties/CHANGELOG.md
@@ -1,3 +1,5 @@
+## Rails 5.0.0.beta3 (February 24, 2016) ##
+
 *   Alias `rake` with `rails_command` in the Rails Application Templates API
     following Rails 5 convention of preferring "rails" to "rake" to run tasks.